Preliminary fixes to catalog manager and gateway. Created catalog service. Changed occurrences of component to product. Got rid of search service and search gateway
Manager: Configured to initialize catalog service instead of catalog gateway. Catalog service is invoked to do tasks. Every instance of component is replaced with product.
Service: Created the class. It initializes the catalog gateway
Gateway: Fixed instances of component to product. Still need to implement the sql.
Still need to implement entities for the gateway to manipulate.